# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DevOps Manager 分布式运维管理系统API文档，支持主机管理、任务调度、文件传输等功能",
        "title": "DevOps Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/hosts": {
            "get": {
                "description": "获取系统中所有已准入的主机信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "主机管理"
                ],
                "summary": "获取主机列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HostListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/hosts/register": {
            "post": {
                "description": "注册一个新的主机到系统中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "主机管理"
                ],
                "summary": "注册新主机",
                "parameters": [
                    {
                        "description": "主机信息",
                        "name": "host",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HostRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/hosts/{id}": {
            "get": {
                "description": "根据主机ID获取主机详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "主机管理"
                ],
                "summary": "获取主机详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主机ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.HostInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/pending-hosts": {
            "get": {
                "description": "获取所有等待管理员准入的主机列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "主机管理"
                ],
                "summary": "获取待准入主机列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PendingHostListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/pending-hosts/{id}/approve": {
            "post": {
                "description": "管理员准入一个待准入的主机",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "主机管理"
                ],
                "summary": "准入主机",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主机ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "description": "获取系统中的任务列表，支持分页和筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "获取任务列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务状态筛选",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务名称筛选",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TaskListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的执行任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "创建新任务",
                "parameters": [
                    {
                        "description": "任务信息",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.TaskResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}/commands": {
            "get": {
                "description": "获取指定任务的所有命令执行记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "获取任务命令列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_message": {
                    "type": "string",
                    "example": "参数验证失败"
                },
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.CreateTaskRequest": {
            "type": "object",
            "required": [
                "command",
                "host_ids",
                "name"
            ],
            "properties": {
                "command": {
                    "type": "string",
                    "example": "bash deploy.sh"
                },
                "description": {
                    "type": "string",
                    "example": "在指定主机上执行部署脚本"
                },
                "host_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "agent-host-001",
                        "agent-host-002"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "执行脚本任务"
                },
                "parameters": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "timeout": {
                    "type": "integer",
                    "example": 300
                }
            }
        },
        "models.HostInfoResponse": {
            "type": "object",
            "properties": {
                "hostname": {
                    "type": "string",
                    "example": "web-server-01"
                },
                "id": {
                    "type": "string",
                    "example": "agent-host-001"
                },
                "ip": {
                    "type": "string",
                    "example": "192.168.1.100"
                },
                "last_seen": {
                    "type": "integer",
                    "example": 1640995200
                },
                "os": {
                    "type": "string",
                    "example": "linux"
                },
                "tags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.HostListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.HostInfoResponse"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.HostRegisterRequest": {
            "type": "object",
            "required": [
                "hostname"
            ],
            "properties": {
                "hostname": {
                    "type": "string",
                    "example": "web-server-01"
                },
                "ip": {
                    "type": "string",
                    "example": "192.168.1.100"
                },
                "os": {
                    "type": "string",
                    "example": "linux"
                },
                "tags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.PendingHostListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PendingHostResponse"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.PendingHostResponse": {
            "type": "object",
            "properties": {
                "first_seen": {
                    "type": "integer",
                    "example": 1640995000
                },
                "host_id": {
                    "type": "string",
                    "example": "agent-pending-001"
                },
                "hostname": {
                    "type": "string",
                    "example": "new-server-01"
                },
                "ip": {
                    "type": "string",
                    "example": "192.168.1.200"
                },
                "last_seen": {
                    "type": "integer",
                    "example": 1640995200
                },
                "os": {
                    "type": "string",
                    "example": "linux"
                },
                "tags": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.TaskListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TaskResponse"
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.TaskResponse": {
            "type": "object",
            "properties": {
                "completed_hosts": {
                    "type": "integer",
                    "example": 5
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T09:55:00Z"
                },
                "created_by": {
                    "type": "string",
                    "example": "admin"
                },
                "description": {
                    "type": "string",
                    "example": "部署Web应用到生产环境"
                },
                "failed_hosts": {
                    "type": "integer",
                    "example": 0
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "部署应用"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "task_id": {
                    "type": "string",
                    "example": "task-001"
                },
                "total_hosts": {
                    "type": "integer",
                    "example": 5
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T10:05:00Z"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}