# server/pkg/controller/http_task_controller.go
package controller

import (
	"github.com/gin-gonic/gin"
)

// HTTPTaskController 任务 HTTP 控制器
type HTTPTaskController struct {
	// taskService *service.TaskService // TODO: 实现 TaskService
}

// NewHTTPTaskController 创建新的任务 HTTP 控制器
func NewHTTPTaskController() *HTTPTaskController {
	return &HTTPTaskController{
		// taskService: service.GetTaskService(), // TODO: 实现 TaskService
	}
}

// RegisterTaskHTTPRoutes 注册任务相关 HTTP 路由
func RegisterTaskHTTPRoutes(r *gin.Engine) {
	controller := NewHTTPTaskController()

	api := r.Group("/api/v1")
	{
		// 任务管理
		api.POST("/tasks", controller.CreateTask)
		api.GET("/tasks", controller.GetTasks)
		api.GET("/tasks/:id", controller.GetTask)
		api.PUT("/tasks/:id", controller.UpdateTask)
		api.DELETE("/tasks/:id", controller.DeleteTask)

		// 任务执行
		api.POST("/tasks/:id/start", controller.StartTask)
		api.POST("/tasks/:id/stop", controller.StopTask)
		api.POST("/tasks/:id/cancel", controller.CancelTask)

		// 任务状态
		api.GET("/tasks/:id/status", controller.GetTaskStatus)
		api.GET("/tasks/:id/progress", controller.GetTaskProgress)
		api.GET("/tasks/:id/logs", controller.GetTaskLogs)

		// 任务主机管理
		api.POST("/tasks/:id/hosts", controller.AddTaskHosts)
		api.DELETE("/tasks/:id/hosts/:host_id", controller.RemoveTaskHost)
		api.GET("/tasks/:id/hosts", controller.GetTaskHosts)

		// 任务命令管理
		api.POST("/tasks/:id/commands", controller.AddTaskCommand)
		api.GET("/tasks/:id/commands", controller.GetTaskCommands)
		api.DELETE("/tasks/:id/commands/:command_id", controller.RemoveTaskCommand)
	}
}

// CreateTask 创建任务
// @Summary      创建新任务
// @Description  创建一个新的执行任务
// @Tags         任务管理
// @Accept       json
// @Produce      json
// @Param        task  body      models.CreateTaskRequest  true  "任务信息"
// @Success      200   {object}  models.APIResponse{data=models.TaskResponse}
// @Failure      400   {object}  models.APIResponse
// @Failure      500   {object}  models.APIResponse
// @Router       /tasks [post]
func (tc *HTTPTaskController) CreateTask(c *gin.Context) {
	// TODO: 实现创建任务逻辑
	SendSuccessResponse(c, gin.H{"message": "Task creation not implemented yet"})
}

// GetTasks 获取任务列表
// @Summary      获取任务列表
// @Description  获取系统中的任务列表，支持分页和筛选
// @Tags         任务管理
// @Accept       json
// @Produce      json
// @Param        page    query     int     false  "页码"        default(1)
// @Param        size    query     int     false  "每页数量"     default(20)
// @Param        status  query     string  false  "任务状态筛选"
// @Param        name    query     string  false  "任务名称筛选"
// @Success      200     {object}  models.TaskListResponse
// @Failure      500     {object}  models.APIResponse
// @Router       /tasks [get]
func (tc *HTTPTaskController) GetTasks(c *gin.Context) {
	// TODO: 实现获取任务列表逻辑
	SendSuccessResponse(c, []gin.H{})
}

// GetTask 获取单个任务
func (tc *HTTPTaskController) GetTask(c *gin.Context) {
	// TODO: 实现获取单个任务逻辑
}

// UpdateTask 更新任务
func (tc *HTTPTaskController) UpdateTask(c *gin.Context) {
	// TODO: 实现更新任务逻辑
}

// DeleteTask 删除任务
func (tc *HTTPTaskController) DeleteTask(c *gin.Context) {
	// TODO: 实现删除任务逻辑
}

// StartTask 启动任务
func (tc *HTTPTaskController) StartTask(c *gin.Context) {
	// TODO: 实现启动任务逻辑
}

// StopTask 停止任务
func (tc *HTTPTaskController) StopTask(c *gin.Context) {
	// TODO: 实现停止任务逻辑
}

// CancelTask 取消任务
func (tc *HTTPTaskController) CancelTask(c *gin.Context) {
	// TODO: 实现取消任务逻辑
}

// GetTaskStatus 获取任务状态
func (tc *HTTPTaskController) GetTaskStatus(c *gin.Context) {
	// TODO: 实现获取任务状态逻辑
}

// GetTaskProgress 获取任务进度
func (tc *HTTPTaskController) GetTaskProgress(c *gin.Context) {
	// TODO: 实现获取任务进度逻辑
}

// GetTaskLogs 获取任务日志
func (tc *HTTPTaskController) GetTaskLogs(c *gin.Context) {
	// TODO: 实现获取任务日志逻辑
}

// AddTaskHosts 添加任务主机
func (tc *HTTPTaskController) AddTaskHosts(c *gin.Context) {
	// TODO: 实现添加任务主机逻辑
}

// RemoveTaskHost 移除任务主机
func (tc *HTTPTaskController) RemoveTaskHost(c *gin.Context) {
	// TODO: 实现移除任务主机逻辑
}

// GetTaskHosts 获取任务主机列表
func (tc *HTTPTaskController) GetTaskHosts(c *gin.Context) {
	// TODO: 实现获取任务主机列表逻辑
}

// AddTaskCommand 添加任务命令
func (tc *HTTPTaskController) AddTaskCommand(c *gin.Context) {
	// TODO: 实现添加任务命令逻辑
}

// GetTaskCommands 获取任务命令列表
// @Summary      获取任务命令列表
// @Description  获取指定任务的所有命令执行记录
// @Tags         任务管理
// @Accept       json
// @Produce      json
// @Param        id   path      string  true  "任务ID"
// @Success      200  {object}  models.APIResponse
// @Failure      404  {object}  models.APIResponse
// @Failure      500  {object}  models.APIResponse
// @Router       /tasks/{id}/commands [get]
func (tc *HTTPTaskController) GetTaskCommands(c *gin.Context) {
	// TODO: 实现获取任务命令列表逻辑
	SendSuccessResponse(c, []gin.H{})
}

// RemoveTaskCommand 移除任务命令
func (tc *HTTPTaskController) RemoveTaskCommand(c *gin.Context) {
	// TODO: 实现移除任务命令逻辑
}